+++ src/proyectoprimerbimestre/PantallaPrincipal.java
@@ -31,7 +31,7 @@ public class PantallaPrincipal extends javax.swing.JFrame {
                 txtTipoCuenta3.setText(""); txtSaldoCuenta3.setText(""); txtNumeroCuenta3.setText("");
                 txtTipoCuenta2.setText(""); txtSaldoCuenta2.setText(""); txtNumeroCuenta2.setText("");
                 txtTipoCuenta1.setText(String.valueOf(cliente.getTipoDeCuenta(0))); 
-                txtSaldoCuenta1.setText("$"+String.valueOf(frmt.format((long)cliente.getSaldo(0))));
+                txtSaldoCuenta1.setText("$"+String.valueOf(frmt.format(cliente.getSaldo(0))));
                 txtNumeroCuenta1.setText("#"+String.valueOf(cliente.getNumeroDeCuenta(0)));
                 btnInfoCuenta4.setVisible(false);
                 btnInfoCuenta3.setVisible(false);
@@ -42,10 +42,10 @@ public class PantallaPrincipal extends javax.swing.JFrame {
                 txtTipoCuenta4.setText(""); txtSaldoCuenta4.setText(""); txtNumeroCuenta4.setText("");
                 txtTipoCuenta3.setText(""); txtSaldoCuenta3.setText(""); txtNumeroCuenta3.setText("");
                 txtTipoCuenta1.setText(String.valueOf(cliente.getTipoDeCuenta(0))); 
-                txtSaldoCuenta1.setText("$"+String.valueOf(frmt.format((long)cliente.getSaldo(0))));
+                txtSaldoCuenta1.setText("$"+String.valueOf(frmt.format(cliente.getSaldo(0))));
                 txtNumeroCuenta1.setText("#"+String.valueOf(cliente.getNumeroDeCuenta(0)));
                 txtTipoCuenta2.setText(String.valueOf(cliente.getTipoDeCuenta(1))); 
-                txtSaldoCuenta2.setText("$"+String.valueOf(frmt.format((long) cliente.getSaldo(1))));
+                txtSaldoCuenta2.setText("$"+String.valueOf(frmt.format( cliente.getSaldo(1))));
                 txtNumeroCuenta2.setText("#"+String.valueOf(cliente.getNumeroDeCuenta(1)));
                 btnInfoCuenta4.setVisible(false);
                 btnInfoCuenta3.setVisible(false);
@@ -54,29 +54,29 @@ public class PantallaPrincipal extends javax.swing.JFrame {
             case 3:
                 txtTipoCuenta4.setText(""); txtSaldoCuenta4.setText(""); txtNumeroCuenta4.setText("");
                 txtTipoCuenta1.setText(String.valueOf(cliente.getTipoDeCuenta(0))); 
-                txtSaldoCuenta1.setText("$"+String.valueOf(frmt.format((long)cliente.getSaldo(0))));
+                txtSaldoCuenta1.setText("$"+String.valueOf(frmt.format(cliente.getSaldo(0))));
                 txtNumeroCuenta1.setText("#"+String.valueOf(cliente.getNumeroDeCuenta(0)));
                 txtTipoCuenta2.setText(String.valueOf(cliente.getTipoDeCuenta(1))); 
-                txtSaldoCuenta2.setText("$"+String.valueOf(frmt.format((long)cliente.getSaldo(1))));
+                txtSaldoCuenta2.setText("$"+String.valueOf(frmt.format(cliente.getSaldo(1))));
                 txtNumeroCuenta2.setText("#"+String.valueOf(cliente.getNumeroDeCuenta(1)));
                 txtTipoCuenta3.setText(String.valueOf(cliente.getTipoDeCuenta(2))); 
-                txtSaldoCuenta3.setText("$"+String.valueOf(frmt.format((long)cliente.getSaldo(2))));
+                txtSaldoCuenta3.setText("$"+String.valueOf(frmt.format(cliente.getSaldo(2))));
                 txtNumeroCuenta3.setText("#"+String.valueOf(cliente.getNumeroDeCuenta(2)));
                 btnInfoCuenta4.setVisible(false);
                
                 break;
             default:
                 txtTipoCuenta1.setText(String.valueOf(cliente.getTipoDeCuenta(0))); 
-                txtSaldoCuenta1.setText("$"+String.valueOf(frmt.format((long)cliente.getSaldo(0))));
+                txtSaldoCuenta1.setText("$"+String.valueOf(frmt.format(cliente.getSaldo(0))));
                 txtNumeroCuenta1.setText("#"+String.valueOf(cliente.getNumeroDeCuenta(0)));
                 txtTipoCuenta2.setText(String.valueOf(cliente.getTipoDeCuenta(1))); 
-                txtSaldoCuenta2.setText("$"+String.valueOf(frmt.format((long)cliente.getSaldo(1))));
+                txtSaldoCuenta2.setText("$"+String.valueOf(frmt.format(cliente.getSaldo(1))));
                 txtNumeroCuenta2.setText("#"+String.valueOf(cliente.getNumeroDeCuenta(1)));
                 txtTipoCuenta3.setText(String.valueOf(cliente.getTipoDeCuenta(2))); 
-                txtSaldoCuenta3.setText("$"+String.valueOf(frmt.format((long)cliente.getSaldo(2))));
+                txtSaldoCuenta3.setText("$"+String.valueOf(frmt.format(cliente.getSaldo(2))));
                 txtNumeroCuenta3.setText("#"+String.valueOf(cliente.getNumeroDeCuenta(2)));
                 txtTipoCuenta4.setText(String.valueOf(cliente.getTipoDeCuenta(3))); 
-                txtSaldoCuenta4.setText("$"+String.valueOf(frmt.format((long)cliente.getSaldo(3))));
+                txtSaldoCuenta4.setText("$"+String.valueOf(frmt.format(cliente.getSaldo(3))));
                 txtNumeroCuenta4.setText("#"+String.valueOf(cliente.getNumeroDeCuenta(3)));
                 
                 break;
